snapshot value finished
