+++ api/services/snapshotvalue/store_getsnapshotvalues.go
@@ -3,10 +3,10 @@ package snapshotvalue
 import (
 	"context"
 	"errors"
+	"fmt"
 
 	"github.com/MicroFish91/portfolio-instruments-api/api/constants"
 	"github.com/MicroFish91/portfolio-instruments-api/api/types"
-	"github.com/jackc/pgx/v5"
 )
 
 func (s *PostgresSnapshotValueStore) GetSnapshotValues(ctx context.Context, snapId, userId int) ([]types.SnapshotValue, error) {
@@ -15,15 +15,15 @@ func (s *PostgresSnapshotValueStore) GetSnapshotValues(ctx context.Context, snap
 
 	rows, err := s.db.Query(
 		c,
-		`
+		fmt.Sprintf(`
 			select
-				*
+				%s
 			from
 				snapshots_values
 			where
 				snap_id = $1
 				and user_id = $2
-		`,
+		`, snapshotValueColumns),
 		snapId,
 		userId,
 	)
@@ -42,29 +42,4 @@ func (s *PostgresSnapshotValueStore) GetSnapshotValues(ctx context.Context, snap
 	}
 
 	return snapshotValues, nil
-}
-
-func (s *PostgresSnapshotValueStore) parseRowsIntoSnapshotValues(rows pgx.Rows) ([]types.SnapshotValue, error) {
-	var svs []types.SnapshotValue
-	for rows.Next() {
-		var sv types.SnapshotValue
-		err := rows.Scan(
-			&sv.Snap_val_id,
-			&sv.Snap_id,
-			&sv.Account_id,
-			&sv.Holding_id,
-			&sv.Total,
-			&sv.Skip_rebalance,
-			&sv.User_id,
-			&sv.Created_at,
-			&sv.Updated_at,
-		)
-
-		if err != nil {
-			return []types.SnapshotValue{}, err
-		}
-		svs = append(svs, sv)
-	}
-
-	return svs, nil
 }
